11 created reducers and root reducer, init redux
Created two reducers, one for project and one for authorization. Then created a root reducer that takes those two in and creates a new object. Imported this root reducer into index.js, and set it as the reducer for our createStore function. Wrapped our app in the Provider (from react-redux) and passed this store over.