14 Imported Thunk, inserted into flow
A very difficult one. Add thunk (Middleware) to index.js and called applyMiddleware. Imported connect from react-redux into CreateProject, and passed mapDispatchToProps as the SECOND argument (first is always mapStateToProps?). This adds a dispatch function called createProject to our props (so that we can pass it down to our component). All this prop function does is take our fucntion and pass it into a dispatch call of our createProject action creator (function). Next we'll add the async code to that action creator.