27 Added some route guards
Fairly straightforward. Pulled the auth from state.firebase.auth and added it to some state components. Those components check if the user is logged in. If it is a page for users, and they're not signed in, we use a <Redirect> to send them to sign in page. If its NOT for users, and they ARE signed in, we do the same thing to Redirect them to the home page.