28 Signing up from website
Using similar logic as our signin page, we can now sign up from the signup page. Added a signUp action that adds a new user to firebase and uses that UID (from the response) to create a new item in the user collection (where we'll store other info later). Also added an error item similar to the signin